src/requirements.txt
pytsk3
pyyaml